+++ lua/github-theme/config.lua
@@ -1,70 +1,41 @@
-local types = require('github-theme.types')
+local vim_config = require('github-theme.config.vim_config')
+local default_config = require('github-theme.config.default')
 
 ---@class gt.Config
+---@field schema gt.ConfigSchema
 local config = {}
 
--- shim vim for kitty and other generators
-vim = vim or { g = {}, o = {} }
-local key_prefix = 'github_'
+---Accessing global config using vim.g.github_* will help in autocompletion.
 
-local function opt(key, default)
-  key = key_prefix .. key
-  if vim.g[key] == nil then
-    vim.g[key] = default
-    return default
-  else
-    if vim.g[key] == 1 then
-      return true
-    elseif vim.g[key] == 0 then
-      return false
-    else
-      return vim.g[key]
-    end
-  end
-end
-
----@type gt.ConfigSchema
 config.schema = {
-  colors = opt('colors', {}),
-  comment_style = opt('comment_style', types.gt.HighlightStyle.Italic),
-  dark_float = opt('dark_float', false),
-  dark_sidebar = opt('dark_sidebar', true),
-  dev = opt('dev', false),
-  function_style = opt('function_style', types.gt.HighlightStyle.None),
-  hide_end_of_buffer = opt('hide_end_of_buffer', true),
-  hide_inactive_statusline = opt('hide_inactive_statusline', true),
-  keyword_style = opt('keyword_style', types.gt.HighlightStyle.Italic),
-  msg_area_style = opt('msg_area_style', types.gt.HighlightStyle.None),
-  overrides = opt('overrides', function()
-    return {}
-  end),
-  sidebars = opt('sidebars', {}),
-  theme_style = opt('theme_style', types.gt.ThemeStyle.Dark),
-  transform_colors = false,
-  transparent = opt('transparent', false),
-  variable_style = opt('variable_style', types.gt.HighlightStyle.None),
+  colors = vim_config.get(vim.g.github_colors, default_config.colors),
+  comment_style = vim_config.get(vim.g.github_comment_style, default_config.comment_style),
+  dark_float = vim_config.get(vim.g.github_dark_float, default_config.dark_float),
+  dark_sidebar = vim_config.get(vim.g.github_dark_sidebar, default_config.dark_sidebar),
+  dev = vim_config.get(vim.g.github_dev, default_config.dev),
+  function_style = vim_config.get(vim.g.github_function_style, default_config.function_style),
+  hide_end_of_buffer = vim_config.get(vim.g.github_hide_end_of_buffer, default_config.hide_end_of_buffer),
+  hide_inactive_statusline = vim_config.get(vim.g.github_hide_inactive_statusline, default_config.hide_inactive_statusline),
+  keyword_style = vim_config.get(vim.g.github_keyword_style, default_config.keyword_style),
+  msg_area_style = vim_config.get(vim.g.github_msg_area_style, default_config.msg_area_style),
+  overrides = vim_config.get(vim.g.github_overrides, default_config.overrides),
+  sidebars = vim_config.get(vim.g.github_sidebars, default_config.sidebars),
+  theme_style = vim_config.get(vim.g.github_theme_style, default_config.theme_style),
+  transparent = vim_config.get(vim.g.github_transparent, default_config.transparent),
+  variable_style = vim_config.get(vim.g.github_variable_style, default_config.variable_style),
 }
 
 ---Override user's configuration
 ---@param user_config gt.ConfigSchema
 config.apply_configuration = function(user_config)
-  --  `set background=light` for these themes
-  local light_background = { light = true, light_default = true }
   for key, value in pairs(user_config) do
     if value ~= nil then
       if config.schema[key] ~= nil then
         -- override value
         config.schema[key] = value
-        vim.g[key_prefix .. key] = value
-
-        -- background set
-        if key == 'theme_style' then
-          if light_background[config.schema.theme_style] then
-            vim.o.background = 'light'
-          end
-        end
+        vim_config.set(key, value)
       else
-        error('projekt0n/github-nvim-theme: Config ' .. key .. ' does not exist') -- luacheck: ignore
+        error('config ' .. key .. ' does not exist') -- luacheck: ignore
       end
     end
   end